!sensors: change magnetic field units
This commit changes magnetic field units in libsensors client api from
mGauss to 0.1 nT for increased sensitivity.

+++ sensors/client-include/libsensors/client.h
@@ -105,9 +105,9 @@ typedef struct {
 /* Magnetometer data */
 typedef struct {
 	uint32_t devId;
-	int16_t magX; /* value in 1E-7 [T] */
-	int16_t magY; /* value in 1E-7 [T] */
-	int16_t magZ; /* value in 1E-7 [T] */
+	int32_t magX; /* value in 0.1 [nT] */
+	int32_t magY; /* value in 0.1 [nT] */
+	int32_t magZ; /* value in 0.1 [nT] */
 	uint8_t reserved[2];
 } mag_data_t;
